backend/requirements.txt
fastapi
uvicorn
mediapipe
opencv-python-headless
numpy
python-multipart
google-generativeai
python-dotenv
requests
